move activation/deactivation to mixins
